requirements.txt
Django==1.8.4
Markdown==2.6.2
django-filter==0.10.0
djangorestframework==3.1.3
pytest==2.8.7
bokeh==0.11.1
requests==2.9.1
PyYAML>=3.11
Sphinx>=1.3.1